% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bnbHelperGof}
\alias{bnbHelperGof}
\title{bnbHelperGof}
\usage{
bnbHelperGof(ancest, children, dat, G, withinAgg, aggType, bs, intercept)
}
\arguments{
\item{ancest}{indices of ancestor in the dat matrix}

\item{children}{indices of children in the dat matrix}

\item{dat}{n x p matrix with observations}

\item{G}{an n x k x |ancest| array of where each variable (corresponding to slice) has k test functions}

\item{withinAgg}{indicates which norm to take when combining test statistics within
a variable but across test functions}

\item{aggType}{indicates which norm to take when combining test statistics across variables}

\item{bs}{number of bootstrap draws}

\item{intercept}{should an additional intercept be included?}
}
\value{
pval: p-value for each child in children
}
\description{
This function is a helper for the branchAndBound function
}
