% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnb.R
\name{brandAndBound}
\alias{brandAndBound}
\title{Confidence sets for Causal Discovery}
\usage{
brandAndBound(Y, K = c(3), bs = 200, aggType = 1, alpha = 0.05,
  pValueAgg = "tippett", intercept = 1, verbose = T)
}
\arguments{
\item{Y}{an n by p matrix with the observations}

\item{K}{a vector containing integer values greater than 2 which will be used for the test statistic}

\item{bs}{the number of bootstrap resamples for the null distribution}

\item{aggType}{the aggregation used for the test statistic
#' \itemize{
\item "inf": \eqn{\max_{j \prec v} \vert (Y_j^K)^T \eta_v \vert} 
\item "1" \eqn{\sum_{j \prec v} \vert (Y_j^K)^T \eta_v \vert}
\item "2" \eqn{\sum_{j \prec v} \vert (Y_j^K)^T \eta_v \vert}
}}

\item{pValueAgg}{the procedure for aggregating p-values 
#' \itemize{
\item "tippet": \eqn{T = \min p_{\theta(v)})} which follows a Beta(1, p-1) distribution
\item "fisher": \eqn{T = -2 \sum_{\theta(v) > 1} \log( p_{\theta(v)})} which follows a chi-squared with p-1 degrees of freedom
}}

\item{intercept}{1 indicates that an intercept should be used; 0 indicates no intercept is used}

\item{verbose}{T or F. Indicates whether to print updates to console}
}
\value{
a vector of p-values corresponding to test
}
\description{
Branch and bound for an exhaustive search
}
