% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnb_anm.R
\name{brandAndBound_anm}
\alias{brandAndBound_anm}
\title{Confidence sets for Causal Discovery}
\usage{
brandAndBound_anm(
  Y,
  G,
  bs = 200,
  aggType = 3,
  alpha = 0.05,
  pValueAgg = "tippett",
  intercept = 1,
  verbose = T,
  basis = "bspline",
  K = 5
)
}
\arguments{
\item{Y}{an n by p matrix with the observations}

\item{G}{a 3d array containing the evaluated test functions. G[i, j, u] is h_j(Y_{u,i})}

\item{bs}{the number of bootstrap resamples for the null distribution}

\item{aggType}{the aggregation used for the test statistic
#' \itemize{
\item "1" \eqn{\sum_{u \prec pr(v), j} \vert h_j(Y_u)^T \eta_v \vert}
\item "2" \eqn{\sum_{u \prec pr(v), j} \vert h_j(Y_u)^T \eta_v \vert^2}
\item "3" or "inf": \eqn{\max_{u \prec pr(v), j} \vert h_j(Y_u)^T \eta_v \vert} 
}}

\item{alpha}{the level of the confidence set}

\item{pValueAgg}{the procedure for aggregating p-values 
#' \itemize{
\item "tippet": \eqn{T = \min p_{\theta(v)})} which follows a Beta(1, p-1) distribution
\item "fisher": \eqn{T = -2 \sum_{\theta(v) > 1} \log( p_{\theta(v)})} which follows a chi-squared with p-1 degrees of freedom
}}

\item{verbose}{T or F. Indicates whether to print updates to console}

\item{basis}{the basis used to model f_v. Choices are "bspline" or "poly"}

\item{K}{either the df for bsplines or the number of polynomial terms}
}
\value{
a vector of p-values corresponding to test
}
\description{
Branch and bound procedure for producing a confidence set of causal orderings
Assumes an additive noise model where the basis is either bsplines or polynomial
}
