% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsProposal.R
\name{bsProp}
\alias{bsProp}
\title{Confidence sets for Causal Discovery}
\usage{
bsProp(Y, K = c(2, 3), propBsCutoff = 1000, numProposals = 100,
  recaptureCutoff = 0.1, proposalBs, bs = 200, aggType = 1,
  intercept = T)
}
\arguments{
\item{Y}{an n by p matrix with the observations}

\item{K}{a vector containing integer values greater than 2 which will be used for the test statistic}

\item{numProposals}{The number of proposals to test for the confidence set}

\item{proposalBs}{the size of the bootstrap sample used to generate the proposals}

\item{bs}{the number of bootstrap resamples for the null distribution when testing any propposal}

\item{aggType}{the aggregation used for the test statistic
#' \itemize{
\item "inf": L-inf 
\item "1" L-1
\item "2" L-2 
\item "3" Computes all of the above
}}

\item{intercept}{T indicates that an intercept should be used; 0 indicates no intercept is used}
}
\value{
a vector of p-values corresponding to test
}
\description{
Test whether a specific ordering is consistent with the data
}
