% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_modSelect.R
\name{ci_modSelect}
\alias{ci_modSelect}
\title{Confidence sets for Causal Discovery}
\usage{
ci_modSelect(tab, treat, outc, effectType = "total", alpha = 0.05, Y)
}
\arguments{
\item{tab}{a matrix with p columns (corresponding to variables) and each row is an ordering}

\item{effectType}{either 'total' or 'direct'}

\item{alpha}{the confidence level (conditional on the selected graph)}

\item{treatment}{index of treatment of interest}

\item{outcome}{index of outcome of interest}
}
\value{
ci is the interval
length is the length
}
\description{
Take a set of orderings and get the confidence interval for a specific
causal effect which accounts for model uncertainty
}
