% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTest.R
\name{exhaustiveTest}
\alias{exhaustiveTest}
\title{Confidence sets for Causal Discovery}
\usage{
exhaustiveTest(Y, K = c(2, 3), aggType = "all", bs = 200,
  intercept = 1)
}
\arguments{
\item{Y}{an n by p matrix with the observations}

\item{K}{a vector containing integer values greater than 2 which will be used for the test statistic}

\item{aggType}{the aggregation used for the test statistic
#' \itemize{
\item "inf": \eqn{\max_{j \prec v} \vert (Y_j^K)^T \eta_v \vert} 
\item "1" \eqn{\sum_{j \prec v} \vert (Y_j^K)^T \eta_v \vert}
\item "2" \eqn{\sum_{j \prec v} \vert (Y_j^K)^T \eta_v \vert}
\item "3" \eqn{\sum_{j \prec v} \vert (Y_j^K)^T \eta_v \vert}
}}

\item{bs}{the number of bootstrap resamples for the null distribution}

\item{intercept}{1 indicates that an intercept should be used; 0 indicates no intercept is used}
}
\value{
a matrix containing p-values corresponding to various tests for all permutations of 1,..., p
}
\description{
Tests all total orderings to see if they are consistent with the data
}
