% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDAG.R
\name{rDAG}
\alias{rDAG}
\title{Sample data from an additive noise model}
\usage{
rDAG(
  p,
  n,
  parent_prob = 1/3,
  lowScale = 1,
  highScale = 1,
  lowEdge = 0.3,
  highEdge = 1,
  edgeVar = NULL,
  intercept = F,
  dist = "gauss",
  uniqueTop = T,
  BInput = NULL,
  scalesInput = NULL,
  intInput = NULL,
  posAndNeg = T
)
}
\arguments{
\item{p}{number of variables}

\item{n}{number of observations}

\item{parent_prob}{the probability of an edge between any two nodes}

\item{lowScale}{lower bound on variance of error terms}

\item{highScale}{upper bound on variance of error terms}

\item{lowEdge}{lower bound on edgeWeights}

\item{highEdge}{upper bound on edgeWeights}

\item{edgeVar}{if null, defaults to uniform draws. If specified, this is the variance of the edgeweights which are drawn from gamma distribution}

\item{intercept}{draw random intercept or leave mean 0}

\item{dist}{the distribution of the error terms. Choices are: "gauss", "unif", "lognorm", "gamma", "weibull", "laplace", "mixed"}

\item{uniqueTop}{Whether to enforce a unique topological ordering so that u -> u+1 for all u}

\item{BInput}{a matrix of edges which can be passed in instead of randomly drawing edges}

\item{scalesInput}{a vector of error variances which can be passed in instead of randomly drawing}

\item{intInput}{a vector of intercepts which can be passed in instead of randomly drawing}

\item{posAndNeg}{whether edges should be both positive and negative or only positive}

\item{AdjMat}{an adjacency matrix which can be passed in to specify the edges instead of randomly selecting a graph}
}
\value{
#' \itemize{
\item B the matrix of edge weights
\item Y the n x p data
\item errs the error realizations
\item scales the variance of the errors
\item mu the intercept terms
}
}
\description{
Sample data from an additive noise model
}
