% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDAG_anm.R
\name{rDAG_anm}
\alias{rDAG_anm}
\title{Confidence sets for Causal Discovery}
\usage{
rDAG_anm(
  p,
  n,
  parent_prob = 1/3,
  lowScale = 1/5,
  highScale = sqrt(2)/5,
  dist = "gauss",
  uniqueTop = T,
  AdjMat = NULL,
  funcType = "GP",
  h = 1,
  sigma = 1,
  M = 50,
  noParentMult = 5
)
}
\arguments{
\item{p}{number of variables}

\item{n}{number of observations}

\item{parent_prob}{the probability of an edge between any two nodes}

\item{lowScale}{lower bound on variance of error terms}

\item{highScale}{upper bound on variance of error terms}

\item{dist}{the distribution of the error terms. Choices are: "gauss", "unif", "lognorm", "gamma", "weibull", "laplace", "mixed"}

\item{uniqueTop}{Whether to enforce a unique topological ordering so that u -> u+1 for all u}

\item{AdjMat}{an adjacency matrix which can be passed in to specify the edges instead of randomly selecting a graph}

\item{funcType:}{Either "GP", "poly" "fourier", "cam", or "tanh"#' @param highScale upper bound on variance of error terms}
}
\value{
#' \itemize{
\item B the p x p adjacency matrix
\item Y the n x p data
\item errs the error realizations
\item scales the variance of the errors
}
}
\description{
Sample data from an additive noise model
}
