% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTest.R
\name{testOrdering}
\alias{testOrdering}
\title{Confidence sets for Causal Discovery}
\usage{
testOrdering(Y, ord, K = c(2, 3), bs = 500, aggType = 1,
  intercept = T)
}
\arguments{
\item{Y}{an n by p matrix with the observations}

\item{ord}{a total ordering of 1, ..., p which will be tested}

\item{K}{a vector containing integer values greater than 2 which will be used for the test statistic}

\item{bs}{the number of bootstrap resamples for the null distribution}

\item{aggType}{the aggregation used for the test statistic
#' \itemize{
\item "inf": L-inf 
\item "1" L-1
\item "2" L-2 
\item "3" Computes all of the above
}}

\item{intercept}{T indicates that an intercept should be used; 0 indicates no intercept is used}
}
\value{
a vector of p-values corresponding to test
}
\description{
Test whether a specific ordering is consistent with the data
}
