% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFit.R
\name{senSen2014}
\alias{senSen2014}
\title{Confidence sets for Causal Discovery}
\usage{
senSen2014(
  X,
  Y,
  bs = 200,
  statType = "hsic",
  K = c(2, 3),
  cutoff = NULL,
  sampleSplit = F,
  kernel = "gaussian"
)
}
\arguments{
\item{X}{an n by p matrix with the covariates}

\item{Y}{an n vector with the observed outcome variables}

\item{bs}{the number of bootstrap resamples for the null distribution}

\item{statType}{#' \itemize{
\item "hsic": uses dHSIC
\item "moment": uses moment based statistic with degree K
\item "dcor": uses distance correlation 
}}

\item{K}{degree of moment for moment based statistic}
}
\value{
a p-value testing the hypothesis
}
\description{
Goodness of fit test from Sen and Sen (2014)
}
